# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sijil API
  version: 0.1.0
  description: Multi-tenant practice management API for Saudi law firms.
servers:
  - url: http://localhost:3001
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, email, password]
              properties:
                tenantId: { type: string, format: uuid }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Tokens and user profile
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New tokens
  /users:
    get:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: List users (Partner)
      responses:
        '200': { description: List of tenant users }
    post:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: Create user (Partner)
      responses:
        '201': { description: User created }
  /settings/tenant:
    get:
      tags: [Settings]
      security: [{ bearerAuth: [] }]
      summary: Get tenant settings
      responses:
        '200': { description: Tenant settings }
    patch:
      tags: [Settings]
      security: [{ bearerAuth: [] }]
      summary: Update tenant settings (Partner)
      responses:
        '200': { description: Tenant settings updated }
  /clients:
    get:
      tags: [Clients]
      security: [{ bearerAuth: [] }]
      summary: List clients
      responses:
        '200': { description: Paged clients }
    post:
      tags: [Clients]
      security: [{ bearerAuth: [] }]
      summary: Create client
      responses:
        '201': { description: Client created }
  /matters:
    get:
      tags: [Matters]
      security: [{ bearerAuth: [] }]
      summary: List matters
      responses:
        '200': { description: Paged matters }
    post:
      tags: [Matters]
      security: [{ bearerAuth: [] }]
      summary: Create matter
      responses:
        '201': { description: Matter created }
  /documents:
    get:
      tags: [Documents]
      security: [{ bearerAuth: [] }]
      summary: List documents
      responses:
        '200': { description: Paged documents }
    post:
      tags: [Documents]
      security: [{ bearerAuth: [] }]
      summary: Create document + first version + upload URL
      responses:
        '201': { description: Document and presigned upload URL }
  /documents/{id}/download-url:
    get:
      tags: [Documents]
      security: [{ bearerAuth: [] }]
      summary: Get signed download URL
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Presigned download URL }
  /documents/{id}/share:
    post:
      tags: [Documents]
      security: [{ bearerAuth: [] }]
      summary: Create expiring share token
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Public share URL }
  /documents/share/{token}:
    get:
      tags: [Documents]
      summary: Public share download
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200': { description: One-time signed URL response }
  /tasks:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: List tasks
      responses:
        '200': { description: Paged tasks }
    post:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: Create task with optional reminder
      responses:
        '201': { description: Task created }
  /billing/quotes:
    get:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: List quotes
      responses:
        '200': { description: Paged quotes }
    post:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: Create quote
      responses:
        '201': { description: Quote created }
  /billing/quotes/{id}/convert:
    post:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: Convert quote to invoice
      responses:
        '200': { description: Invoice created }
  /billing/invoices:
    get:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: List invoices
      responses:
        '200': { description: Paged invoices }
  /billing/invoices/{id}/pay:
    patch:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: Mark invoice paid
      responses:
        '200': { description: Invoice updated }
  /billing/invoices/{id}/pdf:
    get:
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      summary: Export invoice PDF
      responses:
        '200': { description: PDF stream }
  /dashboard/widgets:
    get:
      tags: [Dashboard]
      security: [{ bearerAuth: [] }]
      summary: Dashboard widgets
      responses:
        '200': { description: Aggregated widgets }
  /audit:
    get:
      tags: [Audit]
      security: [{ bearerAuth: [] }]
      summary: Tenant audit log
      responses:
        '200': { description: Paged audit events }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT